Default step time verify
