+++ app/imports/ui/components/actions/includes/create/create.js
@@ -5,11 +5,11 @@ import invoke from 'lodash.invoke';
 import { ActionPlanOptions } from '/imports/api/constants.js';
 import { insert } from '/imports/api/actions/methods.js';
 import { Actions } from '/imports/api/actions/actions.js';
-import { getTzTargetDate, setModalError, inspire } from '/imports/api/helpers.js';
+import { getTzTargetDate, getWorkflowDefaultStepDate, setModalError, inspire } from '/imports/api/helpers.js';
 import { WorkItems } from '/imports/api/work-items/work-items.js';
 
 Template.Actions_Create.viewmodel({
-  mixin: ['workInbox', 'organization', 'router', 'getChildrenData', 'workflow'],
+  mixin: ['workInbox', 'organization', 'router', 'getChildrenData'],
   type: '',
   title: '',
   ownerId: Meteor.userId(),
@@ -19,7 +19,7 @@ Template.Actions_Create.viewmodel({
     const linkedToVM = this.child('Actions_LinkedTo_Edit');
     const linkedTo = linkedToVM && linkedToVM.linkedTo() || [];
 
-    return this.getWorkflowDefaultStepDate({ organization, linkedTo });
+    return getWorkflowDefaultStepDate({ organization, linkedTo });
   },
   toBeCompletedBy: Meteor.userId(),
   verificationTargetDate: '',
