+++ app/imports/ui/components/includes/subcards/actions/edit/action/create/create.js
@@ -2,9 +2,10 @@ import { Template } from 'meteor/templating';
 
 import { ActionPlanOptions } from '/imports/api/constants.js';
 import { insert } from '/imports/api/actions/methods.js';
+import { getWorkflowDefaultStepDate } from '/imports/api/helpers.js';
 
 Template.Actions_CreateSubcard.viewmodel({
-  mixin: ['getChildrenData', 'organization', 'workflow'],
+  mixin: ['getChildrenData', 'organization'],
   type: '',
   title: '',
   ownerId: Meteor.userId(),
@@ -14,7 +15,7 @@ Template.Actions_CreateSubcard.viewmodel({
     const linkedToVM = this.child('Actions_LinkedTo_Edit');
     const linkedTo = linkedToVM && linkedToVM.linkedTo() || [];
 
-    return this.getWorkflowDefaultStepDate({ organization, linkedTo });
+    return getWorkflowDefaultStepDate({ organization, linkedTo });
   },
   toBeCompletedBy: Meteor.userId(),
   verificationTargetDate: '',
